Cottage homework
git-svn-id: https://adminer.svn.sourceforge.net/svnroot/adminer/trunk@77 7c3ca157-0c34-0410-bff1-cbf682f78f5c

+++ create.inc.php
@@ -1,7 +1,7 @@
 <?php
 $types = types();
 $unsigned = array("", "unsigned", "zerofill", "unsigned zerofill");
-if ($_POST && !$_POST["add"]) {
+if ($_POST && !$error && !$_POST["add"]) {
 	if ($_POST["drop"]) {
 		$query = "DROP TABLE " . idf_escape($_GET["create"]);
 		$message = lang('Table has been dropped.');
@@ -28,15 +28,19 @@ if ($_POST && !$_POST["add"]) {
 					. idf_escape($field["field"]) . " $field[type]"
 					. ($field["length"] ? "($field[length])" : "")
 					. (preg_match('~int|float|double|decimal~', $field["type"]) && in_array($field["unsigned"], $unsigned) ? " $field[unsigned]" : "")
-					. (preg_match('~char|text~', $field["type"]) && $field["collation"] ? " COLLATE '" . mysql_real_escape_string($field["collation"]) . "'" : "")
+					. (preg_match('~char|text|enum|set~', $field["type"]) && $field["collation"] ? " COLLATE '" . mysql_real_escape_string($field["collation"]) . "'" : "")
 					. ($field["null"] ? "" : " NOT NULL")
 					. ($key == $_POST["auto_increment"] ? " AUTO_INCREMENT$auto_increment_index" : "")
+					. " COMMENT '" . mysql_real_escape_string($field["comment"]) . "'"
 				;
 			} elseif (strlen($field["orig"])) {
 				$fields[] = "DROP " . idf_escape($field["orig"]);
 			}
 		}
-		$status = ($_POST["Engine"] ? " ENGINE='" . mysql_real_escape_string($_POST["Engine"]) . "'" : "") . ($_POST["Collation"] ? " COLLATE '" . mysql_real_escape_string($_POST["Collation"]) . "'" : "");
+		$status = ($_POST["Engine"] ? " ENGINE='" . mysql_real_escape_string($_POST["Engine"]) . "'" : "")
+			. ($_POST["Collation"] ? " COLLATE '" . mysql_real_escape_string($_POST["Collation"]) . "'" : "")
+			. " COMMENT='" . mysql_real_escape_string($_POST["Comment"]) . "'"
+		;
 		if (strlen($_GET["create"])) {
 			$query = "ALTER TABLE " . idf_escape($_GET["create"]) . " " . implode(", ", $fields) . ", RENAME TO " . idf_escape($_POST["name"]) . ", $status";
 			$message = lang('Table has been altered.');
@@ -57,6 +61,9 @@ if ($_POST) {
 		echo "<p class='error'>" . lang('Unable to operate table') . ": " . htmlspecialchars($error) . "</p>\n";
 	}
 	$row = $_POST;
+	if ($row["auto_increment"]) {
+		$row["fields"][$row["auto_increment"]]["auto_increment"] = true;
+	}
 } elseif (strlen($_GET["create"])) {
 	$row = mysql_fetch_assoc(mysql_query("SHOW TABLE STATUS LIKE '" . mysql_real_escape_string($_GET["create"]) . "'"));
 	$row["name"] = $_GET["create"];
@@ -64,7 +71,9 @@ if ($_POST) {
 } else {
 	$row = array("fields" => array());
 }
-//! default, comments
+if (!$_POST || $_POST["add"]) {
+	$row["fields"][] = array();
+}
 $collations = collations();
 ?>
 <form action="" method="post" id="form">
@@ -74,46 +83,53 @@ $collations = collations();
 <select name="Collation"><option value="">(<?php echo lang('collation'); ?>)</option><?php echo optionlist($collations, $row["Collation"], "not_vals"); ?></select>
 </p>
 <table border="0" cellspacing="0" cellpadding="2">
-<thead><tr><th><?php echo lang('Name'); ?></th><td><?php echo lang('Type'); ?></td><td><?php echo lang('Length'); ?></td><td><?php echo lang('Options'); ?></td><td><?php echo lang('NULL'); ?></td><td><input type="radio" name="auto_increment" value="" /><?php echo lang('Auto-increment'); ?></td></tr></thead>
+<thead><tr><th><?php echo lang('Name'); ?></th><td><?php echo lang('Type'); ?></td><td><?php echo lang('Length'); ?></td><td><?php echo lang('Options'); ?></td><td><?php echo lang('NULL'); ?></td><td><input type="radio" name="auto_increment" value="" /><?php echo lang('Auto Increment'); ?></td><td id="comment-0"><?php echo lang('Comment'); ?></td></tr></thead>
 <?php
 $i=1;
+$column_comments = false;
 foreach ($row["fields"] as $field) {
-	if (strlen($field["field"]) || strlen($field["orig"])) {
-		?>
+	?>
 <tr>
 <th><input type="hidden" name="fields[<?php echo $i; ?>][orig]" value="<?php echo htmlspecialchars($field[($_POST ? "orig" : "field")]); ?>" /><input name="fields[<?php echo $i; ?>][field]" value="<?php echo htmlspecialchars($field["field"]); ?>" maxlength="64" /></th>
 <td><select name="fields[<?php echo $i; ?>][type]" onchange="type_change(this);"><?php echo optionlist(array_keys($types), $field["type"], "not_vals"); ?></select></td>
 <td><input name="fields[<?php echo $i; ?>][length]" value="<?php echo htmlspecialchars($field["length"]); ?>" size="3" /></td>
 <td><select name="fields[<?php echo $i; ?>][collation]"><option value="">(<?php echo lang('collation'); ?>)</option><?php echo optionlist($collations, $field["collation"], "not_vals"); ?></select> <select name="fields[<?php echo $i; ?>][unsigned]"><?php echo optionlist($unsigned, $field["unsigned"], "not_vals"); ?></select></td>
 <td><input type="checkbox" name="fields[<?php echo $i; ?>][null]" value="1"<?php if ($field["null"]) { ?> checked="checked"<?php } ?> /></td>
-<td><input type="radio" name="auto_increment" value="<?php echo $i; ?>"<?php if ($row["auto_increment"] == $i || $field["extra"] == "auto_increment") { ?> checked="checked"<?php } ?> /></td>
+<td><input type="radio" name="auto_increment" value="<?php echo $i; ?>"<?php if ($field["auto_increment"]) { ?> checked="checked"<?php } ?> /></td>
+<td id="comment-<?php echo $i; ?>"><input name="fields[<?php echo $i; ?>][comment]" value="<?php echo htmlspecialchars($field["comment"]); ?>" maxlength="255" /></td>
 </tr>
 <?php
-		$i++;
+	if (strlen($field["comment"])) {
+		$column_comments = true;
 	}
+	$i++;
 }
 //! JavaScript for next rows
 ?>
-<tr>
-<th><input name="fields[<?php echo $i; ?>][field]" maxlength="64" /></th>
-<td><select name="fields[<?php echo $i; ?>][type]" onchange="type_change(this);"><?php echo optionlist(array_keys($types), array(), "not_vals"); ?></select></td>
-<td><input name="fields[<?php echo $i; ?>][length]" size="3" /></td>
-<td><select name="fields[<?php echo $i; ?>][collation]"><option value="">(<?php echo lang('collation'); ?>)</option><?php echo optionlist($collations, array(), "not_vals"); ?></select> <select name="fields[<?php echo $i; ?>][unsigned]"><?php echo optionlist($unsigned, array(), "not_vals"); ?></select></td>
-<td><input type="checkbox" name="fields[<?php echo $i; ?>][null]" value="1" /></td>
-<td><input type="radio" name="auto_increment" value="<?php echo $i; ?>" /></td>
-</tr>
 </table>
+<p><?php echo lang('Comment'); ?>: <input name="Comment" value="<?php echo htmlspecialchars($row["Comment"]); ?>" maxlength="60" />
 <script type="text/javascript">
 function type_change(type) {
 	var name = type.name.substr(0, type.name.length - 6);
-	type.form[name + '[collation]'].style.display = (/char|text/.test(type.form[name + '[type]'].value) ? '' : 'none');
+	type.form[name + '[collation]'].style.display = (/char|text|enum|set/.test(type.form[name + '[type]'].value) ? '' : 'none');
 	type.form[name + '[unsigned]'].style.display = (/int|float|double|decimal/.test(type.form[name + '[type]'].value) ? '' : 'none');
 }
-for (var i=1; <?php echo $i; ?> >= i; i++) {
+for (var i=1; <?php echo $i; ?> > i; i++) {
 	document.getElementById('form')['fields[' + i + '][type]'].onchange();
 }
+
+document.write('<input type="checkbox" id="column_comments"<?php if ($column_comments) { ?> checked="checked"<?php } ?> onclick="column_comments_click(this.checked);" /><label for="column_comments"><?php echo lang('Show column comments'); ?></label>');
+function column_comments_click(checked) {
+	for (var i=0; <?php echo $i; ?> > i; i++) {
+		document.getElementById('comment-' + i).style.display = (checked ? '' : 'none');
+	}
+}
+<?php if (!$column_comments) { ?>column_comments_click(false);<?php } ?>
+
 </script>
+</p>
 <p>
+<input type="hidden" name="token" value="<?php echo $token; ?>" />
 <input type="submit" value="<?php echo lang('Save'); ?>" />
 <?php if (strlen($_GET["create"])) { ?><input type="submit" name="drop" value="<?php echo lang('Drop'); ?>" /><?php } ?>
 </p>
